move panels to services page
