+++ components/Home/welcome.mdx
@@ -6,7 +6,7 @@ Certified in the internationally recognized [Casley-Smith](http://www.casleysmit
 Treatment is highly recommended for:
 - maintaining [wellness](/services/wellness), good circulation & immunity
 - mitigating chronic swelling of [lymphedema](/services/lymphedema)
-- [lipidema](/services/lymphedema)
+- managing [lipidema](/services/lymphedema)
 - preparation [before surgery](/services/surgeries)
 - aiding recovery [post-surgery](/services/surgeries)
