+++ content/services.mdx
@@ -4,5 +4,5 @@ Page content for Services main page
 
 Dynamic Lymphatics provides several services at competetive rates that provide synergistic benefits, relief from pain and stress, improved mobility, strengthened immune system, and reduction of swelling.
 
-More information will be provided here in the future. But for now please call 310-283-9382 to learn more about what services and packages are available.
+For more detail about packages and pricing please call 310-283-9382.
